+++ examples/struct/cocotb/example.py
@@ -17,6 +17,18 @@ class packed_struct_t(avl.Struct):
     multi_bit : avl.Uint32 = avl.Uint32(0)
     state_enum : avl.Enum = avl.Enum("S0", {"S0" : 0, "S1" : 1, "S2" : 2})
 
+class struct_a_b_t(avl.Struct):
+    var_a   : avl.Logic = avl.Logic(0, width = 4)
+    var_b   : avl.Logic = avl.Logic(0, width = 4)
+
+class struct_c_d_t(avl.Struct):
+    var_c   : avl.Logic = avl.Logic(0, width = 4)
+    var_d   : avl.Logic = avl.Logic(0, width = 4)
+
+class struct_a_b_c_d_t(avl.Struct):
+    struct_a_b : struct_a_b_t = struct_a_b_t()
+    struct_c_d : struct_c_d_t = struct_c_d_t()
+
 class example_env(avl.Env):
     def __init__(self, name, parent):
         super().__init__(name, parent)
@@ -76,6 +88,11 @@ class example_env(avl.Env):
             assert self.s0.state_enum == self.s1.state_enum
 
             assert self.s0_copy.multi_bit.value == 200
+
+        await Timer(10, unit="ns")
+
+        # Test nested structs
+        self.test_nested_structs()
         await Timer(10, unit="ns")
 
         # Test the .value shortcut
@@ -99,6 +116,44 @@ class example_env(avl.Env):
         assert self.s0[2:34] == 0xdeadbeef
 
         self.drop_objection()
+
+    def check_struct(self, struct, expected):
+        self.info(f"Struct:{struct}")
+        for field_name, field_type in struct._fields_:
+            field_val = getattr(struct, field_name)
+            if hasattr(field_val, '_fields_'):  # it's a nested struct
+                self.check_struct(field_val, expected)
+            else:  # it's a leaf value
+                self.info(f"  {field_name}: {field_val}")
+                assert field_val == expected[field_name], \
+                    f"variable {field_name} is not correctly assigned. Expected: {hex(expected[field_name])}, Actual: {hex(field_val)}"
+
+    def test_nested_structs(self):
+        _var_a = 0xA
+        _var_b = 0xB
+        _var_c = 0xC
+        _var_d = 0xD
+        init_val = (_var_a << 12) | (_var_b << 8) | (_var_c << 4) | _var_d
+
+        # Create a dict to map field names to expected values
+        expected = {
+            "var_a": _var_a,
+            "var_b": _var_b,
+            "var_c": _var_c,
+            "var_d": _var_d,
+        }
+
+        self.info(f"Testing nested struct with init_val = {hex(init_val)}")
+        full_struct = struct_a_b_c_d_t()
+        #using the from_bits function
+        full_struct.from_bits(init_val)
+        self.info(f"Calling check_struct for inital value")
+        self.check_struct(full_struct, expected)
+        #using the to_bits function
+        self.info(f"Setting a varible using to_bits function")
+        to_bits_var = full_struct.to_bits()
+        self.info(f"Value set is equal to {hex(to_bits_var)}")
+        assert to_bits_var == init_val, f"to_bits is not working. Expected {hex(init_val)}, Actual: {hex(to_bits_var)}"
 
 @cocotb.test
 async def test(dut):